Add getModuleDict utility function to streamline module collection in FlameGraph
